Use FixedLengthStreamingMode which is better supported by web servers. We use the long parameter version of the function, which requires API 19. This version allows to upload files bigger than 2GB

+++ plugin_gvsigol_io/src/main/java/org/gvsig/android/plugin_gvsigol_io/network/NetworkUtilitiesGvsigol.java
@@ -355,11 +355,10 @@ public class NetworkUtilitiesGvsigol {
             setCsrfHeader(manager, conn);
             conn.setDoOutput(true);
             conn.setDoInput(true);
-            //conn.setChunkedStreamingMode(0);
+            
             conn.setRequestProperty("Content-Type", "application/octet-stream");
             conn.setRequestProperty("Content-Length", "" + fileSize);
-            conn.setFixedLengthStreamingMode((int)fileSize);
-
+            conn.setFixedLengthStreamingMode(fileSize);
             conn.connect();
 
             wr = new BufferedOutputStream(conn.getOutputStream());
